feat: moving GetExpiredUptimeChecks to the db package
